requirements.txt
cmake==3.29.2
colorlog==6.8.2
hydra-core==1.3.2
hyperopt==0.2.7
matplotlib==3.8.4
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.6.1
optuna-integration==3.6.0
pandas==2.2.2
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
torch==1.13.0
torch_geometric==2.3.0
torchaudio==0.13.0
torchinfo==1.8.0
torchmetrics==1.3.2
torchsummary==1.5.1
torchvision==0.14.0